add library react select (select role)
