+++ frontend/package.json
@@ -35,6 +35,7 @@
     "react-plotlyjs-ts": "^2.2.2",
     "react-router-dom": "^6.6.2",
     "react-scripts": "4.0.3",
+    "react-select": "^5.7.0",
     "styled-components": "^5.3.0",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
